Refactor SerialId type to u32.

+++ interactive_engine/src/executor/store/src/v2/api/mod.rs
@@ -28,7 +28,7 @@ pub type LabelId = u32;
 pub type PropertyId = u32;
 pub type VertexId = u64;
 pub type EdgeInnerId = u64;
-pub type SerialId = usize;
+pub type SerialId = u32;
 
 pub type Records<T> = Box<dyn Iterator<Item=Result<T>> + Send>;
